Port to GTK4

+++ src/buffer.h
@@ -19,17 +19,19 @@
 #ifndef BUFFER_H_
 #define BUFFER_H_
 
+#include <stddef.h>
+
 #define BUFFER_SIZE (128 * 1024)
 
 void create_buffer(void);
 void delete_buffer(void);
-void put_chars(const char *, unsigned int, gboolean, gboolean);
+void put_chars(const char *, size_t, gboolean, gboolean);
 void clear_buffer(void);
 void write_buffer(void);
-void set_display_func(void (*func)(const char *, unsigned int));
-void unset_display_func(void (*func)(const char *, unsigned int));
+void set_display_func(void (*func)(const char *, size_t));
+void unset_display_func(void (*func)(const char *, size_t));
 void set_clear_func(void (*func)(void));
 void unset_clear_func(void (*func)(void));
-void write_buffer_with_func(void (*func)(const char *, unsigned int));
+void write_buffer_with_func(void (*func)(const char *, size_t));
 
 #endif
